% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_antartida.R
\name{inset_antartida}
\alias{inset_antartida}
\title{Agregar Antártida e Islas del Atlántico Sur a mapas de Argentina}
\usage{
inset_antartida(
  ggplot_map,
  mapping = F,
  capa = 1,
  crs = 22197,
  map_fill = "white",
  map_color = "black",
  map_linewidth = 0.1,
  box_color = "black",
  box_linewidth = 0.1,
  x = 0.15,
  y = 0.05,
  scale = 0.3
)
}
\arguments{
\item{ggplot_map}{Objeto de tipo ggplot que contiene un mapa de Argentina a nivel provincial}

\item{mapping}{Especifica si los polígonos de AIAS debe rellenarse con el mismo color que Tierra del Fuego e Islas Malvinas (aplica al parámetro fill de la capa). Por default es FALSE}

\item{capa}{En caso de que mapping sea TRUE, número de capa del objeto ggplot que contiene los datos para buscar el color del fill}

\item{crs}{Proyección de los datos. Por default es 22197 para darle un aspecto esférico.}

\item{map_fill}{Color de relleno de los polígonos (cuando mapping es FALSE). Por default es "white"}

\item{map_color}{Color de las líneas de los polígonos. Por default es "black"}

\item{map_linewidth}{Grosor de la línea de los polígonos. Por default es 0.1}

\item{box_color}{Color de las líneas de la caja del inset. Por default es "black". Para borrarlo poner NA}

\item{box_linewidth}{Grosor de la línea de la caja del inset. Por default es 0.1}

\item{x}{Posición en la grilla sobre el eje x (se usa cowplot::draw_plot()). Por default es 0.15}

\item{y}{Posición en la grilla sobre el eje y  (se usa cowplot::draw_plot()). Por default es 0.05}

\item{scale}{Factor de escala del inset. Por default es 0.3}
}
\description{
Función para agregar un inset de Antártida e Islas del Atlántico Sur a un mapa de Argentina hecho con ggplot
}
\examples{

 mapa <- ggplot2::ggplot() +
  ggplot2::geom_sf(data = geoAr::get_geo("ARGENTINA","provincia"),
   ggplot2::aes(fill = codprov_censo))

  polArViz::inset_antartida(mapa, mapping = TRUE, capa = 1)

}
