% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_caba.R
\name{inset_caba}
\alias{inset_caba}
\title{Zoom de CABA en un mapa de Argentina}
\usage{
inset_caba(
  ggplot_map,
  x = 500,
  y = 0,
  box_color = "black",
  box_linewidth = 0.5,
  scale = 15,
  radius = 15
)
}
\arguments{
\item{ggplot_map}{Objeto de tipo ggplot que contiene un mapa de Argentina a nivel provincial}

\item{x}{Desplazamiento horizontal del inset respecto del centroide de CABA. Por default es 500.}

\item{y}{Desplazamiento vertical del inset respecto del centroide de CABA. Por default es 0}

\item{box_color}{Color de las líneas del contenedor del inset. Por default es "black". Para borrarlo poner NA}

\item{box_linewidth}{Grosor de la línea del contenedor del inset. Por default es 0.5}

\item{scale}{Factor de escala del polígono de CABA. Por default es 15. En caso de aumentar el mismo, tener en cuenta que puede solaparse con el mapa base y devolver error}

\item{radius}{Radio del círculo contenedor del inset. Por default es 15}
}
\description{
Función para agregar un inset de CABA a un mapa de Argentina hecho con ggplot. La/s capa/s de datos deben contener la variable codprov_censo que viene con geoAr::get_geo() para poder filtrar la información
}
\examples{

 mapa <- ggplot2::ggplot() +
  ggplot2::geom_sf(data = geoAr::get_geo("ARGENTINA","provincia"),
   ggplot2::aes(fill = codprov_censo))

  polArViz::inset_caba(mapa)

}
