% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icg.R
\name{plot_icg}
\alias{plot_icg}
\title{Plot ICG}
\usage{
plot_icg(data)
}
\arguments{
\item{data}{un tibble guardado como objeto en el Enviroment luego de consultar \code{\link[opinAr]{get_icg_raw}} con los microdatos del ICG - UTDT}
}
\value{
Devuelve un objeto de \code{class"gg" "ggplot"} que grafica una serie de tiempo con los valores computados del ICG - UTDT con \code{\link[opinAr]{compute_icg}}
}
\description{
Función que permite una visuaizacion rapida de la serie del Indice de Confianza en el Gobierno (ICG)
}
