% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polArViz.R
\docType{package}
\name{polArViz}
\alias{polArViz}
\title{\code{polArViz} package}
\description{
Caja de Herramientas para visualizar el universo \code{polAr}
See the README on
\href{https://github.com/politicaargentina/polArViz/blob/master/README.md}{Github}
}
